Fix size_t vs int warnings in SWIFT

+++ dunetrigger/triggeralgs/src/TriggerActivityMakerSWIFT.cpp
@@ -134,7 +134,7 @@ namespace triggeralgs {
     std::vector<uint8_t> visited(N, 0);
     float eps2 = eps * eps; //clustering radius
 
-    for (int i = 0; i < N; ++i) {
+    for (size_t i = 0; i < N; ++i) {
       if (visited[i]) continue;
       visited[i] = 1;
 
@@ -142,7 +142,7 @@ namespace triggeralgs {
       const float zi = points[i].z;
       const float ti = points[i].t;
 
-      for (int j = 0; j < N; ++j) {
+      for (size_t j = 0; j < N; ++j) {
         float dz = points[j].z - zi;
         float dt = points[j].t - ti;
         if (dz*dz + dt*dt <= eps2) neigh.push_back(j);
@@ -165,14 +165,14 @@ namespace triggeralgs {
           std::vector<int> neigh2;
           const float zj = points[j].z;
           const float tj = points[j].t;
-          for (int m = 0; m < N; ++m) {
+          for (size_t m = 0; m < N; ++m) {
             float dz = points[m].z - zj;
             float dt = points[m].t - tj;
             if (dz*dz + dt*dt <= eps2) neigh2.push_back(m);
           }
 
           if (neigh2.size() >= static_cast<size_t>(min_samples)) {
-            for (int m : neigh2) {
+            for (size_t m : neigh2) {
               if (std::find(neigh.begin(), neigh.end(), m) == neigh.end())
                 neigh.push_back(m);
             }
@@ -188,7 +188,7 @@ namespace triggeralgs {
 
     // once clusters are formed, calculate the energies
     std::vector<uint64_t> cluster_sums(cluster_id, 0);
-    for (int i = 0; i < N; ++i) {
+    for (size_t i = 0; i < N; ++i) {
       if (labels[i] >= 0) cluster_sums[labels[i]] += points[i].adc;
     }
     //return dominant cluster energy in window
